requirements.txt
streamlit
numpy 
opencv-python
numpy
opencv-python
Pillow
torch>=1.7
torchvision
tqdm